refactor(infra): structure, primitive arch
