+++ infra/lib.rs
@@ -2,5 +2,5 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-pub mod datastructs;
-pub mod primitives;
+pub mod structure;
+pub mod primitive;
