+++ infra/structure/lists/peekable.rs
